@@ -2,7 +2,7 @@
 [metadata]
 name = scenedetect
 version = attr: scenedetect.__version__
-license = BSD 3-Clause License
+license = BSD-3-Clause
 author = Brandon Castellano
 author_email = brandon248@gmail.com
 description = Video scene cut/shot detection program and Python library.
@@ -21,10 +21,8 @@ classifiers =
     Intended Audience :: Developers
     Intended Audience :: End Users/Desktop
     Intended Audience :: System Administrators
-    License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
@@ -39,7 +37,8 @@ keywords = video computer-vision analysis
 
 [options]
 install_requires =
-    Click
+    # click <8.3.0 is excluded as per https://scenedetect.com/issues/521.
+    click~=8.0,<8.3.0
     numpy
     platformdirs
     tqdm
@@ -50,12 +49,12 @@ packages =
     scenedetect.backends
     scenedetect.detectors
     scenedetect.output
-python_requires = >=3.7
+python_requires = >=3.8
 
 [options.extras_require]
 opencv = opencv-python
 opencv-headless = opencv-python-headless
-pyav = av
+pyav = av>=9.2
 moviepy = moviepy
 
 [options.entry_points]
