@@ -118,8 +118,8 @@ class StatsManager:
         """Register a list of metric keys that will be used by the detector."""
         self._metric_keys = self._metric_keys.union(set(metric_keys))
 
-    # TODO(#507): This interface is difficult to use, we should support the dictionary protocol.
-    # Ideally this should work with Panadas. This could be done with the v0.7 API change.
+    # TODO(https://scenedetect.com/issues/507): We should support the dictionary protocol instead
+    # of using this bespoke interface. It would be useful for Pandas compatibility as well.
     def get_metrics(
         self, timecode: FrameTimecode, metric_keys: ty.Iterable[str]
     ) -> ty.List[ty.Any]:
@@ -190,7 +190,7 @@ class StatsManager:
         logger.info("Writing %d frames to CSV...", len(frame_keys))
         for frame_key in frame_keys:
             csv_writer.writerow(
-                [frame_key.get_frames() + 1, frame_key.get_timecode()]
+                [frame_key.frame_num + 1, frame_key.get_timecode()]
                 + [str(metric) for metric in self.get_metrics(frame_key, metric_keys)]
             )
 
