@@ -21,6 +21,8 @@ or string HH:MM:SS[.nnn]. timecode format.
 """
 
 # Third-Party Library Imports
+from fractions import Fraction
+
 import pytest
 
 # Standard Library Imports
@@ -38,11 +40,11 @@ def test_framerate():
         FrameTimecode(timecode=None, fps=FrameTimecode(timecode=0, fps=None))
     # Test zero FPS/negative.
     with pytest.raises(ValueError):
-        FrameTimecode(timecode=0, fps=0)
+        FrameTimecode(timecode=0, fps=0.0)
     with pytest.raises(ValueError):
-        FrameTimecode(timecode=0, fps=-1)
+        FrameTimecode(timecode=0, fps=-1.0)
     with pytest.raises(ValueError):
-        FrameTimecode(timecode=0, fps=-100)
+        FrameTimecode(timecode=0, fps=-100.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode=0, fps=0.0)
     with pytest.raises(ValueError):
@@ -52,26 +54,28 @@ def test_framerate():
     with pytest.raises(ValueError):
         FrameTimecode(timecode=0, fps=MAX_FPS_DELTA / 2)
     # Test positive framerates.
-    assert FrameTimecode(timecode=0, fps=1).frame_num == 0
-    assert FrameTimecode(timecode=0, fps=MAX_FPS_DELTA).frame_num == 0
-    assert FrameTimecode(timecode=0, fps=10).frame_num == 0
+    assert FrameTimecode(timecode=0, fps=1.0).frame_num == 0
+    assert FrameTimecode(timecode=0, fps=10.0).frame_num == 0
     assert FrameTimecode(timecode=0, fps=MAX_FPS_DELTA * 2).frame_num == 0
-    assert FrameTimecode(timecode=0, fps=1000).frame_num == 0
     assert FrameTimecode(timecode=0, fps=1000.0).frame_num == 0
+    assert FrameTimecode(timecode=0, fps=1000.0).frame_num == 0
+    # Reject framerates too small for equality testing or potential divide by zero situations.
+    with pytest.raises(ValueError):
+        assert FrameTimecode(timecode=0, fps=MAX_FPS_DELTA).frame_num == 0
 
 
 def test_timecode_numeric():
     """Test FrameTimecode constructor argument "timecode" with numeric arguments."""
     with pytest.raises(ValueError):
-        FrameTimecode(timecode=-1, fps=1)
+        FrameTimecode(timecode=-1, fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode=-1.0, fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode=-0.1, fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode=-1.0 / 1000, fps=1.0)
-    assert FrameTimecode(timecode=0, fps=1).frame_num == 0
-    assert FrameTimecode(timecode=1, fps=1).frame_num == 1
+    assert FrameTimecode(timecode=0, fps=1.0).frame_num == 0
+    assert FrameTimecode(timecode=1, fps=1.0).frame_num == 1
     assert FrameTimecode(timecode=0.0, fps=1.0).frame_num == 0
     assert FrameTimecode(timecode=1.0, fps=1.0).frame_num == 1
 
@@ -80,13 +84,13 @@ def test_timecode_string():
     """Test FrameTimecode constructor argument "timecode" with string arguments."""
     # Invalid strings:
     with pytest.raises(ValueError):
-        FrameTimecode(timecode="-1", fps=1)
+        FrameTimecode(timecode="-1", fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode="-1.0", fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode="-0.1", fps=1.0)
     with pytest.raises(ValueError):
-        FrameTimecode(timecode="1.9x", fps=1)
+        FrameTimecode(timecode="1.9x", fps=1.0)
     with pytest.raises(ValueError):
         FrameTimecode(timecode="1x", fps=1.0)
     with pytest.raises(ValueError):
@@ -95,21 +99,21 @@ def test_timecode_string():
         FrameTimecode(timecode="1.0-", fps=1.0)
 
     # Frame number integer [int->str] ('%d', integer number as string)
-    assert FrameTimecode(timecode="0", fps=1).frame_num == 0
-    assert FrameTimecode(timecode="1", fps=1).frame_num == 1
+    assert FrameTimecode(timecode="0", fps=1.0).frame_num == 0
+    assert FrameTimecode(timecode="1", fps=1.0).frame_num == 1
     assert FrameTimecode(timecode="10", fps=1.0).frame_num == 10
 
     # Seconds format [float->str] ('%f', number as string)
-    assert FrameTimecode(timecode="0.0", fps=1).frame_num == 0
-    assert FrameTimecode(timecode="1.0", fps=1).frame_num == 1
+    assert FrameTimecode(timecode="0.0", fps=1.0).frame_num == 0
+    assert FrameTimecode(timecode="1.0", fps=1.0).frame_num == 1
     assert FrameTimecode(timecode="10.0", fps=1.0).frame_num == 10
     assert FrameTimecode(timecode="10.0000000000", fps=1.0).frame_num == 10
     assert FrameTimecode(timecode="10.100", fps=1.0).frame_num == 10
     assert FrameTimecode(timecode="1.100", fps=10.0).frame_num == 11
 
     # Seconds format [float->str] ('%fs', number as string followed by 's' for seconds)
-    assert FrameTimecode(timecode="0s", fps=1).frame_num == 0
-    assert FrameTimecode(timecode="1s", fps=1).frame_num == 1
+    assert FrameTimecode(timecode="0s", fps=1.0).frame_num == 0
+    assert FrameTimecode(timecode="1s", fps=1.0).frame_num == 1
     assert FrameTimecode(timecode="10s", fps=1.0).frame_num == 10
     assert FrameTimecode(timecode="10.0s", fps=1.0).frame_num == 10
     assert FrameTimecode(timecode="10.0000000000s", fps=1.0).frame_num == 10
@@ -117,70 +121,66 @@ def test_timecode_string():
     assert FrameTimecode(timecode="1.100s", fps=10.0).frame_num == 11
 
     # Standard timecode format [timecode->str] ('HH:MM:SS[.nnn]', where [.nnn] is optional)
-    assert FrameTimecode(timecode="00:00:01", fps=1).frame_num == 1
-    assert FrameTimecode(timecode="00:00:01.9999", fps=1).frame_num == 2
-    assert FrameTimecode(timecode="00:00:02.0000", fps=1).frame_num == 2
-    assert FrameTimecode(timecode="00:00:02.0001", fps=1).frame_num == 2
+    assert FrameTimecode(timecode="00:00:01", fps=1.0).frame_num == 1
+    assert FrameTimecode(timecode="00:00:01.9999", fps=1.0).frame_num == 2
+    assert FrameTimecode(timecode="00:00:02.0000", fps=1.0).frame_num == 2
+    assert FrameTimecode(timecode="00:00:02.0001", fps=1.0).frame_num == 2
 
     # MM:SS[.nnn] is also allowed
-    assert FrameTimecode(timecode="00:01", fps=1).frame_num == 1
-    assert FrameTimecode(timecode="00:01.9999", fps=1).frame_num == 2
-    assert FrameTimecode(timecode="00:02.0000", fps=1).frame_num == 2
-    assert FrameTimecode(timecode="00:02.0001", fps=1).frame_num == 2
+    assert FrameTimecode(timecode="00:01", fps=1.0).frame_num == 1
+    assert FrameTimecode(timecode="00:01.9999", fps=1.0).frame_num == 2
+    assert FrameTimecode(timecode="00:02.0000", fps=1.0).frame_num == 2
+    assert FrameTimecode(timecode="00:02.0001", fps=1.0).frame_num == 2
 
     # Conversion edge cases
-    assert FrameTimecode(timecode="00:00:01", fps=10).frame_num == 10
-    assert FrameTimecode(timecode="00:00:00.5", fps=10).frame_num == 5
-    assert FrameTimecode(timecode="00:00:00.100", fps=10).frame_num == 1
-    assert FrameTimecode(timecode="00:00:00.001", fps=1000).frame_num == 1
+    assert FrameTimecode(timecode="00:00:01", fps=10.0).frame_num == 10
+    assert FrameTimecode(timecode="00:00:00.5", fps=10.0).frame_num == 5
+    assert FrameTimecode(timecode="00:00:00.100", fps=10.0).frame_num == 1
+    assert FrameTimecode(timecode="00:00:00.001", fps=1000.0).frame_num == 1
 
-    assert FrameTimecode(timecode="00:00:59.999", fps=1).frame_num == 60
-    assert FrameTimecode(timecode="00:01:00.000", fps=1).frame_num == 60
-    assert FrameTimecode(timecode="00:01:00.001", fps=1).frame_num == 60
+    assert FrameTimecode(timecode="00:00:59.999", fps=1.0).frame_num == 60
+    assert FrameTimecode(timecode="00:01:00.000", fps=1.0).frame_num == 60
+    assert FrameTimecode(timecode="00:01:00.001", fps=1.0).frame_num == 60
 
-    assert FrameTimecode(timecode="00:59:59.999", fps=1).frame_num == 3600
-    assert FrameTimecode(timecode="01:00:00.000", fps=1).frame_num == 3600
-    assert FrameTimecode(timecode="01:00:00.001", fps=1).frame_num == 3600
+    assert FrameTimecode(timecode="00:59:59.999", fps=1.0).frame_num == 3600
+    assert FrameTimecode(timecode="01:00:00.000", fps=1.0).frame_num == 3600
+    assert FrameTimecode(timecode="01:00:00.001", fps=1.0).frame_num == 3600
 
     # Check too many ":" characters (https://github.com/Breakthrough/PySceneDetect/issues/476)
     with pytest.raises(ValueError):
-        FrameTimecode(timecode="01:01:00:00.001", fps=1)
+        FrameTimecode(timecode="01:01:00:00.001", fps=1.0)
 
 
 def test_get_frames():
     """Test FrameTimecode get_frames() method."""
-    assert FrameTimecode(timecode=1, fps=1.0).get_frames(), 1
-    assert FrameTimecode(timecode=1000, fps=60.0).get_frames(), 1000
-    assert FrameTimecode(timecode=1000000000, fps=29.97).get_frames(), 1000000000
+    assert FrameTimecode(timecode=1, fps=1.0).frame_num == 1
+    assert FrameTimecode(timecode=1000, fps=60.0).frame_num == 1000
+    assert FrameTimecode(timecode=1000000000, fps=29.97).frame_num == 1000000000
 
-    assert FrameTimecode(timecode=1.0, fps=1.0).get_frames(), int(1.0 / 1.0)
-    assert FrameTimecode(timecode=1000.0, fps=60.0).get_frames(), int(1000.0 * 60.0)
-    assert FrameTimecode(timecode=1000000000.0, fps=29.97).get_frames(), int(1000000000.0 * 29.97)
+    assert FrameTimecode(timecode=1.0, fps=1.0).frame_num == int(1.0 / 1.0)
+    assert FrameTimecode(timecode=1000.0, fps=60.0).frame_num == int(1000.0 * 60.0)
+    assert FrameTimecode(timecode=1000000000.0, fps=29.97).frame_num == int(1000000000.0 * 29.97)
 
-    assert FrameTimecode(timecode="00:00:02.0000", fps=1).get_frames(), 2
-    assert FrameTimecode(timecode="00:00:00.5", fps=10).get_frames(), 5
-    assert FrameTimecode(timecode="00:00:01", fps=10).get_frames(), 10
-    assert FrameTimecode(timecode="00:01:00.000", fps=1).get_frames(), 60
+    assert FrameTimecode(timecode="00:00:02.0000", fps=1.0).frame_num == 2
+    assert FrameTimecode(timecode="00:00:00.5", fps=10.0).frame_num == 5
+    assert FrameTimecode(timecode="00:00:01", fps=10.0).frame_num == 10
+    assert FrameTimecode(timecode="00:01:00.000", fps=1.0).frame_num == 60
 
 
 def test_get_seconds():
     """Test FrameTimecode get_seconds() method."""
-    assert FrameTimecode(timecode=1, fps=1.0).get_seconds(), pytest.approx(1.0 / 1.0)
-    assert FrameTimecode(timecode=1000, fps=60.0).get_seconds(), pytest.approx(1000 / 60.0)
-    assert FrameTimecode(timecode=1000000000, fps=29.97).get_seconds(), pytest.approx(
-        1000000000 / 29.97
-    )
+    assert FrameTimecode(timecode=1, fps=1.0).seconds, pytest.approx(1.0 / 1.0)
+    assert FrameTimecode(timecode=1000, fps=60.0).seconds, pytest.approx(1000 / 60.0)
+    assert FrameTimecode(timecode=1000000000, fps=29.97).seconds, pytest.approx(1000000000 / 29.97)
 
-    assert FrameTimecode(timecode=1.0, fps=1.0).get_seconds(), pytest.approx(1.0)
-    assert FrameTimecode(timecode=1000.0, fps=60.0).get_seconds(), pytest.approx(1000.0)
-    assert FrameTimecode(timecode=1000000000.0, fps=29.97).get_seconds(), pytest.approx(
-        1000000000.0
-    )
+    assert FrameTimecode(timecode=1.0, fps=1.0).seconds, pytest.approx(1.0)
+    assert FrameTimecode(timecode=1000.0, fps=60.0).seconds, pytest.approx(1000.0)
+    assert FrameTimecode(timecode=1000000000.0, fps=29.97).seconds, pytest.approx(1000000000.0)
 
-    assert FrameTimecode(timecode="00:00:02.0000", fps=1).get_seconds(), pytest.approx(2.0)
-    assert FrameTimecode(timecode="00:00:00.5", fps=10).get_seconds(), pytest.approx(0.5)
-    assert FrameTimecode(timecode="00:00:01", fps=10).get_seconds(), pytest.approx(1.0)
-    assert FrameTimecode(timecode="00:01:00.000", fps=1).get_seconds(), pytest.approx(60.0)
+    assert FrameTimecode(timecode="00:00:02.0000", fps=1.0).seconds, pytest.approx(2.0)
+    assert FrameTimecode(timecode="00:00:00.5", fps=10.0).seconds, pytest.approx(0.5)
+    assert FrameTimecode(timecode="00:00:01", fps=10.0).seconds, pytest.approx(1.0)
+    assert FrameTimecode(timecode="00:01:00.000", fps=1.0).seconds, pytest.approx(60.0)
 
 
 def test_get_timecode():
@@ -189,10 +189,17 @@ def test_get_timecode():
     assert FrameTimecode(timecode=60.117, fps=60.0).get_timecode() == "00:01:00.117"
     assert FrameTimecode(timecode=3600.234, fps=29.97).get_timecode() == "01:00:00.234"
 
-    assert FrameTimecode(timecode="00:00:02.0000", fps=1).get_timecode() == "00:00:02.000"
-    assert FrameTimecode(timecode="00:00:00.5", fps=10).get_timecode() == "00:00:00.500"
-    assert FrameTimecode(timecode="00:00:01.501", fps=10).get_timecode() == "00:00:01.500"
-    assert FrameTimecode(timecode="00:01:00.000", fps=1).get_timecode() == "00:01:00.000"
+    assert FrameTimecode(timecode="00:00:02.0000", fps=1.0).get_timecode() == "00:00:02.000"
+    assert FrameTimecode(timecode="00:00:00.5", fps=10.0).get_timecode() == "00:00:00.500"
+    # If a value is provided in seconds, we store that value internally now.
+    assert (
+        FrameTimecode(timecode="00:00:01.501", fps=10.0).get_timecode(nearest_frame=False)
+        == "00:00:01.501"
+    )
+    assert (
+        FrameTimecode(timecode="00:00:01.501", fps=10.0).get_timecode(nearest_frame=True)
+        == "00:00:01.500"
+    )
 
 
 def test_equality():
@@ -201,14 +208,10 @@ def test_equality():
     assert x == x
     assert x == FrameTimecode(timecode=1.0, fps=10.0)
     assert x == FrameTimecode(timecode=1.0, fps=10.0)
+    assert x == FrameTimecode(timecode=1.0, fps=Fraction(10, 1))
     assert x != FrameTimecode(timecode=10.0, fps=10.0)
     assert x != FrameTimecode(timecode=10.0, fps=10.0)
-    # Comparing FrameTimecodes with different framerates raises a TypeError.
-    with pytest.raises(TypeError):
-        assert x == FrameTimecode(timecode=1.0, fps=100.0)
-    with pytest.raises(TypeError):
-        assert x == FrameTimecode(timecode=1.0, fps=10.1)
-
+    assert x != FrameTimecode(timecode=10.0, fps=Fraction(100, 10))
     assert x == FrameTimecode(x)
     assert x == FrameTimecode(1.0, x)
     assert x == FrameTimecode(10, x)
@@ -234,13 +237,8 @@ def test_equality():
     with pytest.raises(TypeError):
         assert x == {0: 0}
 
-    assert FrameTimecode(timecode="00:00:00.5", fps=10) == "00:00:00.500"
-    assert FrameTimecode(timecode="00:00:01.500", fps=10) == "00:00:01.500"
-    assert FrameTimecode(timecode="00:00:01.500", fps=10) == "00:00:01.501"
-    assert FrameTimecode(timecode="00:00:01.500", fps=10) == "00:00:01.502"
-    assert FrameTimecode(timecode="00:00:01.500", fps=10) == "00:00:01.508"
-    assert FrameTimecode(timecode="00:00:01.500", fps=10) == "00:00:01.509"
-    assert FrameTimecode(timecode="00:00:01.519", fps=10) == "00:00:01.510"
+    assert FrameTimecode(timecode="00:00:00.5", fps=10.0) == "00:00:00.500"
+    assert FrameTimecode(timecode="00:00:01.500", fps=10.0) == "00:00:01.500"
 
 
 def test_addition():
@@ -248,13 +246,10 @@ def test_addition():
     x = FrameTimecode(timecode=1.0, fps=10.0)
     assert x + 1 == FrameTimecode(timecode=1.1, fps=10.0)
     assert x + 1 == FrameTimecode(1.1, x)
+    assert x + 10 == "00:00:02.000", str(x + 10)
     assert x + 10 == 20
     assert x + 10 == 2.0
-
     assert x + 10 == "00:00:02.000"
-
-    with pytest.raises(TypeError):
-        assert FrameTimecode("00:00:02.000", fps=20.0) == x + 10
 
 
 def test_subtraction():
@@ -263,26 +258,24 @@ def test_subtraction():
     assert (x - 1) == FrameTimecode(timecode=0.9, fps=10.0)
     assert x - 2 == FrameTimecode(0.8, x)
     assert x - 10 == FrameTimecode(0.0, x)
-    # TODO(v1.0): Allow negative values
+    # TODO(v1.0): Allow negative values. For now we clamp.
     assert x - 11 == FrameTimecode(0.0, x)
     assert x - 100 == FrameTimecode(0.0, x)
-
     assert x - 1.0 == FrameTimecode(0.0, x)
     assert x - 100.0 == FrameTimecode(0.0, x)
-
     assert x - 1 == FrameTimecode(timecode=0.9, fps=10.0)
-
-    with pytest.raises(TypeError):
-        assert FrameTimecode("00:00:02.000", fps=20.0) == x - 10
+    assert FrameTimecode("00:00:00.000", fps=20.0) == x - 10
 
 
-@pytest.mark.parametrize("frame_num,fps", [(1, 1), (61, 14), (29, 25), (126, 24000 / 1001.0)])
+@pytest.mark.parametrize(
+    "frame_num,fps", [(1, 1.0), (61, 14.0), (29, 25.0), (126, Fraction(24000, 1001))]
+)
 def test_identity(frame_num, fps):
     """Test FrameTimecode values, when used in init return the same values"""
     frame_time_code = FrameTimecode(frame_num, fps=fps)
     assert FrameTimecode(frame_time_code) == frame_time_code
-    assert FrameTimecode(frame_time_code.get_frames(), fps=fps) == frame_time_code
-    assert FrameTimecode(frame_time_code.get_seconds(), fps=fps) == frame_time_code
+    assert FrameTimecode(frame_time_code.frame_num, fps=fps) == frame_time_code
+    assert FrameTimecode(frame_time_code.seconds, fps=fps) == frame_time_code
     assert FrameTimecode(frame_time_code.get_timecode(), fps=fps) == frame_time_code
 
 
@@ -304,3 +297,12 @@ def test_precision():
     assert FrameTimecode(990, fps).get_timecode(precision=1, use_rounding=False) == "00:00:00.9"
     assert FrameTimecode(990, fps).get_timecode(precision=0, use_rounding=True) == "00:00:01"
     assert FrameTimecode(990, fps).get_timecode(precision=0, use_rounding=False) == "00:00:00"
+
+
+# TODO(v0.8): Remove this test during the removal of `scenedetect.scene_detector`.
+def test_deprecated_timecode_module_emits_warning_on_import():
+    FRAME_TIMECODE_WARNING = (
+        "The `frame_timecode` submodule is deprecated, import from the base package instead."
+    )
+    with pytest.warns(DeprecationWarning, match=FRAME_TIMECODE_WARNING):
+        from scenedetect.frame_timecode import FrameTimecode as _
